move kv store to pkg
because it is a reusable component that other projects can benefit too.
