+++ app/nexus/internal/state/base/locks.go
@@ -5,8 +5,9 @@
 package base
 
 import (
+	"github.com/spiffe/spike/app/nexus/internal/env"
 	"github.com/spiffe/spike/app/nexus/internal/state/entity/data"
-	"github.com/spiffe/spike/app/nexus/internal/state/store"
+	"github.com/spiffe/spike/pkg/store"
 	"sync"
 )
 
@@ -17,7 +18,7 @@ var (
 	adminToken   string
 	adminTokenMu sync.RWMutex
 
-	kv   = store.NewKV()
+	kv   = store.NewKV(store.KVConfig{MaxSecretVersions: env.MaxSecretVersions()})
 	kvMu sync.RWMutex
 
 	adminCredentials   data.Credentials
