+++ pkg/store/delete.go
