+++ pkg/store/delete_test.go
