+++ pkg/store/err.go
